Moving ServiceDataSourceModule, removing unused import from ClientMapper
